Fix the page glitch when scrolling through videos
Changing innerHTML makes it collapse the elements and re-fill them.
Instead it is better to pre-populate the videos with a/img elements and
then in the event handler fill them in with the links on each click.

This prevents the layout from collapsing, and makes the page stop
jumping up when you are at the bottom on scroll through the thumbnails
from camera #3