queue: Use h264 encoder instead of libvpx
In tests it is 3-4x faster than webm and I am having problems with it
creating new videos in a reasonable amount of time. Current versions of
firefox, including on the iPhone, appear to play the video just fine.

Also scale it down to 1280x960, which speeds up the encoding even more.