Add .details.json cache to events
There's no reason to keep re-calculating the thumbnail, video url, etc.
after the first time so this adds storage of the info in a json file
named .details.json for faster loading.

If there is an error reading the file it will be recreated.

If a parameter effecting the data (eg. thumbnail offset) is changed the
cache can be 'cleared' by deleting all the .details.json files.

This also changes the event limit value to return all events if the
limit is set to 0.