Add a page to display all events for a selected camera.
Clicking on a thumbnail now takes you to the new page, with the selected
video playing. Clicking on any of the thumbnails on the page will jump
back to the top and play the selected event.

TODO: Make the playback smaller, and scroll the bottom independent from
the video player at the top.