Start using modern build system configuration
Things have changed since I originally wrote this. It's time to move the
configuration into setup.cfg and the new pyproject.toml files. Simplify
the requirements so that they are only listed in setup.cfg, not in 3
different files.

NOTE: While tox.ini can support embedding the ini in pyproject.toml it
is only as a raw string so there is no point in switching that at this
time.