events: Add expire support to the cache
The cache is kept in memory and on disk. When _check_cache minutes have
elapsed it will check for old data (older than _keep_days) and remove it
from memory and from disk.

Filesystem deletes can be quite expensive so the directory is renamed to
a temporary location and then deleted in a background thread.