events: Fix expiring and deleting old events
The first problem was that I left in clearing the remove dict. This
prevented it from even trying to remove them.

The second big problem was that TemporaryDirectory tries to unlink the
directory when it gets garbage collected, which resulted in the UI
getting stuck unexpectedly and in a hard to debug way (no logs, no
errors on the console). Replaced it with mkdtemp() to fix that.

Also removed the force_expire logic, and just set the time out so that
it wouldn't expire during startup unless startup took took long, even
then it would still respect the counter. Previously it was checking
every single time an event was added to the queue, resulting in a
potentially massive slowdown at startup.