thumbnails: Switch to better thumbnail selection
Previously it was picking the image that was 25% from the start of the
event. This worked, mostly, but I have been tagging all the images with
data about them from motion. They contain the changed pixel count,
coordinates of the box for movement, and the noise level.

The box isn't a great way to tell if it has the 'best' picture because
it can cover a larger area than the actual changes (eg. two separate
active areas are included in the box).

So I now sort based on the changed pixel count and pick the largest one.
So far this seems to work better than the 25% method.