events: Add function to list the events
Use this instead of globbing the filesystem, which can take a noticeable
amount of time depending on the state of the disk cache. Instead do the
glob only at startup to get the initial state of things. All API calls
go through the EventCache list.

This also adds queue_events which will listen on a Pipe (connected to
the queue process) and add new event directories as they are created.