Fix event expire
It was expiring whole days, not just the events within the day that were
old. IIRC I did this because I was having problems with colliding events
in the initial implementation, but this left some events in the cache
that were deleted from disk.

This fixes it to move each individual event to a camera + date specific
delete directory to prevent collision and only expire the events that
are old.