events: Clean up unprocessed events
When strix shuts down or crashes it can leave old event directories
laying around. This is because the queue file is removed before the
event is finished processing in order to prevent long running processes
from being handled more than once.

To fix this I added a check to find all the camera paths that don't
match the hh-mm-ss pattern and to re-create the queue entries for them
so that they will eventually get picked up and processed.

Initially this will also pick up really old events, which will then be
deleted by the cache system, so it is inefficient but it's simple and it
should be much faster once the initial backlog is finished processing.